Plugin now works on TeamCity 8 Moved plugin assembly to maven (deployment through ant still works)
