+++ redmine-server/src/main/resources/buildServerResources/style.css
