--- resources/jar/META-INF/build-server-plugin-redmine.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE beans PUBLIC "-//SPRING//DTD BEAN//EN" "http://www.springframework.org/dtd/spring-beans.dtd">
-
-<beans default-autowire="constructor">
-  <bean id="redmineFetcher" class="com.marcusilgner.redcity.RedmineIssueFetcher">
-    <constructor-arg index="0" ref="ehCacheUtil"/>
-  </bean>
-  <bean id="redmineFactory" class="com.marcusilgner.redcity.RedmineIssueProviderFactory">
-    <constructor-arg index="0" ref="redmineFetcher"/>
-  </bean>
-</beans>
